Rename tag_?cloud to wordcloud.
I think wordcloud is a better name for this as a general purpose
app - there's nothing taggish about any of this really until
it's displayed on the template.
