+++ pombola/ghana/templates/home.html
@@ -9,7 +9,7 @@
 
   <script type = "text/javascript">
     $(function() {
-        $.getJSON("tagcloud/tagcloud/",
+        $.getJSON("wordcloud/wordcloud/",
         {},
         function(data){
             $("#wordcloud").jQCloud(data, {width: 525, height:345, shape: "rectangular"});
