+++ pombola/urls.py
@@ -91,10 +91,10 @@ urlpatterns += patterns('',
     (r'^search/', include('pombola.search.urls')),
 )
 
-# tagcloud
+# wordcloud
 # FIXME - should make this an optional feature
 urlpatterns += patterns('',
-    (r'^tagcloud/', include('tag_cloud.urls')),
+    (r'^wordcloud/', include('wordcloud.urls')),
 )
 
 # feedback
